1. Update Nuget packages  2. Move ContainerCreationService acceptance tests to Acceptance folder
