+++ IronFoundry.Container.Test/Acceptance/ContainerAcceptanceTests.cs
@@ -2,15 +2,12 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using IronFoundry.Container.Utilities;
-using IronFoundry.Warden.Test;
-using NSubstitute;
 using Xunit;
 
 namespace IronFoundry.Container.Acceptance
 {
-    public class ContainerTests : IDisposable
+    public class ContainerAcceptanceTests : IDisposable
     {
         string Container1Handle { get; set; }
         string Container2Handle { get; set; }
@@ -24,7 +21,7 @@ namespace IronFoundry.Container.Acceptance
         IContainer Container1 { get; set; }
         IContainer Container2 { get; set; }
 
-        public ContainerTests()
+        public ContainerAcceptanceTests()
         {
             Container1Handle = GenerateRandomAlphaString();
             Container2Handle = GenerateRandomAlphaString();
@@ -80,8 +77,8 @@ namespace IronFoundry.Container.Acceptance
             Container1.Run(pSpec, io1).WaitForExit();
             Container2.Run(pSpec, io2).WaitForExit();
 
-            string user1 = io1.Output.ToString();
-            string user2 = io2.Output.ToString();
+            var user1 = io1.Output.ToString();
+            var user2 = io2.Output.ToString();
 
             Assert.NotEmpty(user1);
             Assert.NotEmpty(user2);
@@ -179,19 +176,29 @@ namespace IronFoundry.Container.Acceptance
 
         public IContainer CreateContainer(IContainerService containerService, string handle)
         {
-            var bindMounts = new BindMount[]
+            var bindMounts = new[]
             {
-                new BindMount { Access = FileAccess.Read, SourcePath = ReadOnlyBindMountPath, DestinationPath = ReadOnlyBindMountPath },
-                new BindMount { Access = FileAccess.ReadWrite, SourcePath = ReadWriteBindMountPath, DestinationPath = ReadWriteBindMountPath },
+                new BindMount
+                {
+                    Access = FileAccess.Read,
+                    SourcePath = ReadOnlyBindMountPath,
+                    DestinationPath = ReadOnlyBindMountPath
+                },
+                new BindMount
+                {
+                    Access = FileAccess.ReadWrite,
+                    SourcePath = ReadWriteBindMountPath,
+                    DestinationPath = ReadWriteBindMountPath
+                }
             };
 
             var environment = new Dictionary<string, string>
             {
-                { "CONTAINER_HANDLE", handle },
-                { "CONTAINER_ENV1", "ENV1" },
+                {"CONTAINER_HANDLE", handle},
+                {"CONTAINER_ENV1", "ENV1"}
             };
 
-            ContainerSpec spec = new ContainerSpec
+            var spec = new ContainerSpec
             {
                 BindMounts = bindMounts,
                 Environment = environment,
@@ -203,12 +210,12 @@ namespace IronFoundry.Container.Acceptance
             return container;
         }
 
-        static string CreateTempDirectory()
+        private static string CreateTempDirectory()
         {
             string containerBasePath = null;
-            for(int attempt = 0; attempt < 10 && string.IsNullOrWhiteSpace(containerBasePath); attempt++)
+            for (var attempt = 0; attempt < 10 && string.IsNullOrWhiteSpace(containerBasePath); attempt++)
             {
-                string tempPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
+                var tempPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
                 if (!Directory.Exists(tempPath))
                 {
                     containerBasePath = tempPath;
@@ -225,15 +232,15 @@ namespace IronFoundry.Container.Acceptance
             return containerBasePath;
         }
 
-        static string GenerateRandomAlphaString(int length = 4)
+        private static string GenerateRandomAlphaString(int length = 8)
         {
             const string alphabet = "abcdefghijklmnopqrstuvwxyz";
 
-            Random r = RandomFactory.Create();
-            string handle = "";
-            for (int count = 0; count < length; count++)
+            var r = RandomFactory.Create();
+            var handle = "";
+            for (var count = 0; count < length; count++)
             {
-                int chosenCharIndex = r.Next(0, alphabet.Length);
+                var chosenCharIndex = r.Next(0, alphabet.Length);
                 handle += alphabet[chosenCharIndex];
             }
 
@@ -243,8 +250,8 @@ namespace IronFoundry.Container.Acceptance
 
     internal class StringProcessIO : IProcessIO
     {
-        public StringWriter Output = new StringWriter();
         public StringWriter Error = new StringWriter();
+        public StringWriter Output = new StringWriter();
 
         public TextWriter StandardOutput
         {
