+++ IronFoundry.Container.Test/IronFoundry.Container.Test.csproj
@@ -13,7 +13,6 @@
     <AssemblyName>IronFoundry.Container.Test</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
-    <NuGetPackageImportStamp>e00bf8a3</NuGetPackageImportStamp>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -38,10 +37,11 @@
     </Reference>
     <Reference Include="Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
-      <HintPath>..\packages\Newtonsoft.Json.6.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
+      <HintPath>..\packages\Newtonsoft.Json.6.0.8\lib\net45\Newtonsoft.Json.dll</HintPath>
     </Reference>
-    <Reference Include="NSubstitute">
-      <HintPath>..\packages\NSubstitute.1.7.1.0\lib\NET40\NSubstitute.dll</HintPath>
+    <Reference Include="NSubstitute, Version=1.8.1.0, Culture=neutral, PublicKeyToken=92dd2e9066daa5ca, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\packages\NSubstitute.1.8.1.0\lib\net45\NSubstitute.dll</HintPath>
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
@@ -65,7 +65,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Acceptance\ContainerTests.cs" />
+    <Compile Include="Acceptance\ContainerCreationServiceAcceptanceTests.cs" />
+    <Compile Include="Acceptance\ContainerAcceptanceTests.cs" />
     <Compile Include="ConstrainedProcessRunnerTests.cs" />
     <Compile Include="ConstrainedProcessTests.cs" />
     <Compile Include="ContainerServiceTests.cs" />
@@ -81,8 +82,11 @@
     <Compile Include="Messaging\MessagingClientTest.cs" />
     <Compile Include="Messaging\SecureStringJsonConverterTest.cs" />
     <Compile Include="Messaging\TaskExtensions.cs" />
+    <Compile Include="TestSupport\FactAdminRequired.cs" />
     <Compile Include="TestSupport\IFTestHelper.cs" />
     <Compile Include="TestSupport\InputSource.cs" />
+    <Compile Include="TestSupport\RandomFactory.cs" />
+    <Compile Include="TestSupport\TestProcessIO.cs" />
     <Compile Include="Utilities\EffectiveAccessComputerTests.cs" />
     <Compile Include="Utilities\EnvironmentBlockTests.cs" />
     <Compile Include="Utilities\FileSystemEffectiveAccessComputerTests.cs" />
@@ -94,7 +98,6 @@
     <Compile Include="TestSupport\AssertHelper.cs" />
     <Compile Include="TestSupport\NSubstituteExtensions.cs" />
     <Compile Include="TestSupport\TempFile.cs" />
-    <Compile Include="TestSupport\TestProcessIO.cs" />
     <Compile Include="Utilities\FileSystemSecurityDescriptorReaderTests.cs" />
     <Compile Include="Utilities\JobObjectLimitsTests.cs" />
     <Compile Include="Utilities\JobObjectTests.cs" />
@@ -104,8 +107,6 @@
     <Compile Include="ProcessRunnerTests.cs" />
     <Compile Include="ProcessTrackerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestSupport\RandomFactory.cs" />
-    <Compile Include="TestSupport\FactAdminRequired.cs" />
     <Compile Include="Utilities\UndoStackTests.cs" />
   </ItemGroup>
   <ItemGroup>
