+++ IronFoundry.Container.Test/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Newtonsoft.Json" version="6.0.1" targetFramework="net45" />
-  <package id="NSubstitute" version="1.7.1.0" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net45" />
+  <package id="NSubstitute" version="1.8.1.0" targetFramework="net45" />
   <package id="SharpZipLib" version="0.86.0" targetFramework="net45" />
   <package id="xunit" version="2.0.0-rc1-build2826" targetFramework="net45" />
   <package id="xunit.abstractions" version="2.0.0-rc1-build2826" targetFramework="net45" />
