+++ IronFoundry.Warden.IISHost/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="CommandLineParser" version="1.9.71" targetFramework="net45" />
-  <package id="NLog" version="2.1.0" targetFramework="net45" />
+  <package id="NLog" version="3.2.0.0" targetFramework="net45" />
 </packages>
