+++ IronFoundry.Warden.Service/App.config
@@ -26,16 +26,14 @@
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
         <assemblyIdentity name="NLog" publicKeyToken="5120e14c03d0593c" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-2.0.1.0" newVersion="2.0.1.0" />
+        <bindingRedirect oldVersion="0.0.0.0-3.2.0.0" newVersion="3.2.0.0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="Topshelf" publicKeyToken="b800c4cfcdeea87b" culture="neutral" />
         <bindingRedirect oldVersion="0.0.0.0-3.1.122.0" newVersion="3.1.122.0" />
       </dependentAssembly>
       <dependentAssembly>
-        <assemblyIdentity   name="Newtonsoft.Json"
-                              publicKeyToken="30ad4fe6b2a6aeed"
-                              culture="neutral" />
+        <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
         <bindingRedirect oldVersion="4.5.0.0" newVersion="6.0.0.0" />
       </dependentAssembly>
     </assemblyBinding>
