+++ IronFoundry.Warden.Service/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NLog" version="2.1.0" targetFramework="net45" />
+  <package id="NLog" version="3.2.0.0" targetFramework="net45" />
   <package id="protobuf-net" version="2.0.0.668" targetFramework="net45" />
-  <package id="Topshelf" version="3.1.3" targetFramework="net45" />
-  <package id="Topshelf.NLog" version="3.1.3" targetFramework="net45" />
+  <package id="Topshelf" version="3.1.4" targetFramework="net45" />
+  <package id="Topshelf.NLog" version="3.1.4" targetFramework="net45" />
 </packages>
