+++ IronFoundry.Warden/Containers/ContainerHostLauncher.cs
@@ -110,7 +110,7 @@ namespace IronFoundry.Warden.Containers
                     }
                     catch (Exception e)
                     {
-                        log.LogException(LogLevel.Error, e.ToString(), e);
+                        log.Log(LogLevel.Error, e.ToString(), e);
                     }
                     return Task.FromResult(0);
                 });
