+++ IronFoundry.Warden/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Newtonsoft.Json" version="6.0.1" targetFramework="net45" />
-  <package id="NLog" version="2.1.0" targetFramework="net45" />
+  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net45" />
+  <package id="NLog" version="3.2.0.0" targetFramework="net45" />
   <package id="protobuf-net" version="2.0.0.668" targetFramework="net45" />
   <package id="SharpZipLib" version="0.86.0" targetFramework="net45" />
 </packages>
